Add files via upload
In the world with advanced technology, where few confidential tasks are done through the medium of voice there is need for Voice Recognition. There are many chances that may outbreak with improper recognition of one’s voice such as security, authentication, accessibility, convenience and communication. Deep Learning Technique is one of the efficient techniques that will make use of artificial neural networks using Machine Learning Algorithms to detect tasks with provided training data. Here we use Deep Learning Technique to recognize the difference between True and Fake voice considering the 50 neurons that are associated in the hidden layer architecture for the purpose of pattern recognition with 80 is to 20 of percent training and testing data. The extracted features also include the MFCC (Mel-Frequency Cepstral Coefficients) with thirteen coefficients for each frame. Further the performance is evaluated based on different parameters such as accuracy, precision, recall, F1-score, and confusion matrix analysis, approaching to be an effective model to distinguish the True and Fake voice signals in a detailed manner.  

Dataset : For this study, we obtained the audio files from the Kaggle dataset for voice classification, which consists of two main folders: classified audio files and extracted 
coefficients. This dataset was obtained from a publicly available file on Kaggle having Fake Or Real audio files 4GB [https://www.kaggle.com/datasets/birdy654/deep-voice-deepfake-voice-recognition].